// src/store/actions/types.ts
export enum ActionTypes {
    SET_LOADING = "SET_LOADING",
    SET_TITLE = "SET_TITLE",
    LOADING_MORE_GAMES = "LOADING_MORE_GAMES",
    GET_GAME = "GET_GAME",
    GET_GAMES = "GET_GAMES",
    GET_MORE_GAMES = "GET_MORE_GAMES",
    SET_NEXT_API_PAGE = "SET_NEXT_API_PAGE",
    SET_CATEGORY = "SET_CATEGORY",
    SET_ORDER_BY = "SET_ORDER_BY",
    SET_PLATFORM = "SET_PLATFORM",
    SET_GENRE = "SET_GENRE",
    SET_SEARCH_QUERY = "SET_SEARCH_QUERY",
    SEARCH_GAMES = "SEARCH_GAMES",
    LOAD_FAVORITES = "LOAD_FAVORITES",
    LIKE_GAME = "LIKE_GAME",
    DISLIKE_GAME = "DISLIKE_GAME",
    SET_THEME = "SET_THEME"
}